Remove duplicate code
