+++ lib/dataset/util.cpp
@@ -13,6 +13,8 @@
 #include "scipp/variable/variable.h"
 #include "scipp/variable/variable_concept.h"
 
+#include "dataset_operations_common.h"
+
 using namespace scipp::variable;
 namespace scipp {
 
@@ -141,3 +143,29 @@ scipp::index size_of(const Dataset &dataset, const SizeofTag tag) {
   return size;
 }
 } // namespace scipp
+
+namespace scipp::dataset {
+DataArray strip_edges_along(const DataArray &da, const Dim dim) {
+  auto out = da;
+  for (const auto &[name, var] : da.coords())
+    if (core::is_edges(da.dims(), var.dims(), dim))
+      out.coords().erase(name);
+  for (const auto &[name, var] : da.masks())
+    if (core::is_edges(da.dims(), var.dims(), dim))
+      out.masks().erase(name);
+  for (const auto &[name, var] : da.attrs())
+    if (core::is_edges(da.dims(), var.dims(), dim))
+      out.attrs().erase(name);
+  return out;
+}
+
+Dataset strip_edges_along(const Dataset &ds, const Dim dim) {
+  auto out = apply_to_items(
+      ds, [](auto &&...args) { return strip_edges_along(args...); }, dim);
+  for (const auto &[name, var] : ds.coords())
+    if (!core::is_edges(ds.sizes(), var.dims(), dim))
+      out.setCoord(name, var);
+  return out;
+}
+
+} // namespace scipp::dataset
